bug 744442 - Make MINIDUMP_SAVE_PATH work to save minidumps to a known location after processing them. r=jmaher
--HG--
extra : rebase_source : d8a9135b7eb6ae2b6e0e788fca0926b4adc59ada

+++ talos/ttest.py
@@ -53,6 +53,7 @@ import platform
 import os
 import re
 import time
+import shutil
 import sys
 import subprocess
 import utils
@@ -185,7 +186,13 @@ class TTest(object):
                 except:
                     raise talosError("error executing: '%s'" % subprocess.list2cmdline(cmd))
                 nullfd.close()
-            os.remove(dump)
+            dumpSavePath = os.environ.get('MINIDUMP_SAVE_PATH', None)
+            if dumpSavePath:
+                shutil.move(dump, dumpSavePath)
+                utils.noisy("Saved dump as %s" % os.path.join(dumpSavePath,
+                                                              os.path.basename(dump)))
+            else:
+                os.remove(dump)
             found = True
 
         if browser_config['remote'] == True:
